refactor(pane): Phase 4 — lift pane CEF callbacks + wire install_pane_focus_redirect
Payoff phase of the pane modularization split described in
\`docs/specs/SPEC_BROWSER_PANE_MODULARIZATION.md\`. Two things here:

1. **Code motion.** Pane-specific branches of \`AgentMuxHandler\`'s
   \`on_after_created\`, \`on_before_close\`, and \`on_load_end\` move out of
   \`client.rs\` into a new \`agentmux-cef/src/pane/callbacks.rs\`. The
   \`client.rs\` branches become one-line delegations to the new free
   functions. \`is_pane\`-gated inline logic is gone.

2. **Behavior fix.** \`install_pane_focus_redirect\` (from Phase 2) was
   dead code — defined in \`pane::hwnd\` with zero callers. Spec
   \`SPEC_BROWSER_PANE_LIFECYCLE.md\` §5 race #5 flagged this as the cause
   of post-navigation focus-steal regressions: Chromium destroys and
   recreates the \`Chrome_RenderWidgetHostHWND\` on every navigation, so
   the subclass installed at create-time is on a stale HWND by the time
   the user navigates.

   \`pane::callbacks::on_after_created_pane\` now installs the subclass
   at create time, and \`on_load_end_pane\` reinstalls it after every
   navigation so the new widget HWND is always covered.

Files
-----

\`agentmux-cef/src/pane/callbacks.rs\` (new, 90 lines):
- \`on_after_created_pane(state, browser)\` — Z-order raise + install focus subclass.
- \`on_before_close_pane(state, label)\` — drain lifecycle entry.
- \`on_load_end_pane(state, browser)\` — reinstall focus subclass.

\`agentmux-cef/src/pane/mod.rs\`:
- \`pub mod callbacks;\`
- Drop the \`#[allow(unused_imports)]\` on \`install_pane_focus_redirect\`
  (now has a live caller).

\`agentmux-cef/src/pane/hwnd.rs\`:
- Drop the \`#[allow(dead_code)]\` on \`install_pane_focus_redirect\`.

\`agentmux-cef/src/client.rs\`:
- \`on_after_created\`: Z-order + subclass block replaced with call.
- \`on_before_close\`: drain block replaced with call.
- \`on_load_end\`: pane branch now dispatches to \`on_load_end_pane\`
  instead of logging and returning. Main-only work still skipped via
  early return.

Cycle status
------------

The \`browser_panes ↔ client\` direct import cycle was already half-broken
after Phase 2 (\`browser_panes\` no longer imports \`client\`) and fully
broken after Phase 3 (\`CreatePaneTask\` moved to \`pane/creation\`).

With Phase 4, \`client.rs\` no longer has any \`is_pane\`-gated inline
logic either — it only knows that panes exist via \`self.is_pane\` (the
dispatch flag) and by delegating to \`crate::pane::callbacks\`. The
remaining edges: \`pane::creation\` → \`client::{AgentMuxHandler,AgentMuxClient}\`
(one-way, type-only) and \`state::AppState\` → \`browser_panes::BrowserPaneManager\`
(field, not a call cycle).

Testing
-------

\`cargo test\` — 27/27 pass.
\`cargo build\` — 17 warnings, down from 18 (\`install_pane_focus_redirect\`
is no longer dead code).

Behavior change — needs a smoke test
------------------------------------

Unlike Phases 1-3, this phase exercises previously unreachable code
(\`install_pane_focus_redirect\`). If the subclass install has any latent
issue — wrong HWND, pointer bug — it would show up under \`task dev\` or
the portable build.
